requirements.txt
streamlit

pandas

numpy

scipy

matplotlib

seaborn

plotly

google-generativeai

python-dotenv